moved AuthorizationType.java to 'filter' package
